+++ mercury/mercury-library/agent-didcommx/src/main/scala/io/iohk/atala/mercury/MessagingService.scala
@@ -3,42 +3,68 @@ package io.iohk.atala.mercury
 import scala.jdk.CollectionConverters.*
 import zio._
 
+import io.circe._
+import io.circe.Json._
+import io.circe.parser._
+import io.circe.JsonObject
 import io.iohk.atala.mercury.model._
 import io.iohk.atala.mercury.model.error._
+import io.iohk.atala.mercury.protocol.routing._
 import io.iohk.atala.resolvers.DIDResolver
 import org.didcommx.didcomm.common.VerificationMethodType
 import org.didcommx.didcomm.common.VerificationMaterialFormat
 
 type HttpOrDID = String //TODO
 case class ServiceEndpoint(uri: HttpOrDID, accept: Option[Seq[String]], routingKeys: Option[Seq[String]])
+case class MessageAndAddress(msg: Message, url: String)
 
 object MessagingService {
 
-  /** Encrypt and send a Message via HTTP
-    *
-    * TODO Move this method to another model
-    */
-  def send(msg: Message): ZIO[DidComm & DIDResolver & HttpClient, SendMessageError, HttpResponseBody] = {
+  /** Encrypted payload (message) and make the Forward Message */
+  def makeForwardMessage(message: Message, mediator: DidId): URIO[DidComm, ForwardMessage] =
+    for {
+      didCommService <- ZIO.service[DidComm]
+      encrypted <- didCommService.packEncrypted(message, to = message.to.head) // TODO head
+      msg = ForwardMessage(
+        to = mediator,
+        expires_time = None,
+        body = ForwardBody(next = message.to.head), // TODO check msg head
+        attachments = Seq(AttachmentDescriptor.buildJsonAttachment(payload = encrypted.asJson)),
+      )
+    } yield (msg)
+
+  /** Create a Message and any Forward Message as needed */
+  def makeMessage(msg: Message): ZIO[DidComm & DIDResolver & HttpClient, SendMessageError, MessageAndAddress] =
     for {
       didCommService <- ZIO.service[DidComm]
       resolver <- ZIO.service[DIDResolver]
       sendToDID <- msg.to match {
         case Seq() => // TODO support for anonymous message
-          ZIO.fail(new RuntimeException("Missing the destination DID - TODO support for anonymous message"))
+          ZIO.fail(
+            SendMessageError(
+              new RuntimeException("Missing the destination DID - TODO support for anonymous message")
+            )
+          )
         case Seq(value) =>
           ZIO.succeed(value)
         case head +: tail => // TODO support for multiple destinations
-          ZIO.fail(new RuntimeException("TODO multiple destinations"))
+          ZIO.fail(
+            SendMessageError(
+              new RuntimeException("TODO multiple destinations")
+            )
+          )
       }
 
-      encryptedForwardMessage <- didCommService.packEncryptedForward(msg, to = msg.to.head) // TODO head
-      jsonString = encryptedForwardMessage.string
-
-      didCommService <- resolver
+      serviceEndpoint <- resolver
         .didCommServices(sendToDID) /* Seq[DIDCommService] */
+        .catchAll { case ex => ZIO.fail(SendMessageError(ex)) }
         .flatMap {
           case Seq() =>
-            ZIO.fail(new RuntimeException("To send a Message you need a destination")) // TODO ERROR
+            ZIO.fail(
+              SendMessageError(
+                new RuntimeException("To send a Message you need a destination") // TODO ERROR
+              )
+            )
           case Seq(v) =>
             ZIO.succeed(
               ServiceEndpoint(
@@ -57,20 +83,52 @@ object MessagingService {
                 )
               )
         }
-      serviceEndpoint <-
-        if (didCommService.uri.startsWith("did:"))
-          resolver
-            .didCommServices(DidId(didCommService.uri))
-            .map(_.toSeq.head.getServiceEndpoint()) // TODO this is not safe and also need to be recursive
-        else ZIO.succeed(didCommService.uri)
 
-      _ <- Console.printLine("Sending to " + serviceEndpoint)
-      res <- HttpClient.postDIDComm(serviceEndpoint, jsonString)
+      msgToSend <- serviceEndpoint match {
+        case ServiceEndpoint(url, _, None) if url.startsWith("http") =>
+          ZIO.log(s"No Forward Message needed. (send to $url)") *>
+            ZIO.succeed(MessageAndAddress(msg, url))
+        case ServiceEndpoint(url, _, Some(Seq())) if url.startsWith("http") =>
+          ZIO.log(s"No Forward Message needed. (send to $url)") *>
+            ZIO.succeed(MessageAndAddress(msg, url))
+        case ServiceEndpoint(did, _, _) if did.startsWith("did:") =>
+          for {
+            _ <- ZIO.log(s"Make Forward Message for Mediator '$did'")
+            mediator = DidId(did)
+            forwardMessage <- makeForwardMessage(message = msg, mediator = mediator)
+            finalMessage <- makeMessage(forwardMessage.asMessage) // Maybe it needs a double warping
+          } yield finalMessage
+        case ServiceEndpoint(uri, _, Some(routingKeys)) =>
+          ZIO.log(s"RoutingDID: $routingKeys") *> ??? // ZIO.fail(???) // FIXME no support for routingKeys
+        case s @ ServiceEndpoint(_, _, None) =>
+          ZIO.logError(s"Unxpected ServiceEndpoint $s") *> ZIO.fail(
+            SendMessageError(new RuntimeException(s"Unxpected ServiceEndpoint $s"))
+          )
+      }
+    } yield (msgToSend)
+
+    /** Encrypt and send a Message via HTTP
+      *
+      * TODO Move this method to another model
+      */
+  def send(msg: Message): ZIO[DidComm & DIDResolver & HttpClient, SendMessageError, HttpResponseBody] =
+    for {
+      auxFinalMessage <- makeMessage(msg)
+      MessageAndAddress(finalMessage, serviceEndpoint) = auxFinalMessage
+      didCommService <- ZIO.service[DidComm]
+      encryptedMessage <-
+        if (finalMessage.`type` == ForwardMessage.PIURI)
+          didCommService.packEncryptedAnon(msg = finalMessage, to = finalMessage.to.head) // TODO Head
+        else
+          didCommService.packEncrypted(msg = finalMessage, to = finalMessage.to.head) // TODO Head
+
+      _ <- ZIO.log(s"Sending to Message to '$serviceEndpoint'")
+      res <- HttpClient
+        .postDIDComm(url = serviceEndpoint, data = encryptedMessage.string)
+        .catchAll { case ex => ZIO.fail(SendMessageError(ex, Some(encryptedMessage.string))) }
     } yield (res)
-  }.catchAll { case ex =>
-    ZIO.fail(SendMessageError(ex))
-  }
 
+}
 // WIP do not delete!
 //   /** Encrypt and send a Message via HTTP
 //     *
@@ -158,4 +216,3 @@ object MessagingService {
 //       res <- postDIDComm(serviceEndpoint)
 //     } yield (res)
 //   }
-}